+++ app/src/main/java/com/iGap/activities/ActivityMain.java
@@ -1078,7 +1078,7 @@ public class ActivityMain extends ActivityEnhanced
     }
 
     @Override
-    public void onFileDownload(final String token, final int offset, final ProtoFileDownload.FileDownload.Selector selector, final int progress) {
+    public void onFileDownload(final String token, final long offset, final ProtoFileDownload.FileDownload.Selector selector, final int progress) {
         runOnUiThread(new Runnable() {
             @Override
             public void run() {
@@ -1088,7 +1088,7 @@ public class ActivityMain extends ActivityEnhanced
     }
 
     @Override
-    public void onAvatarDownload(final String token, final int offset, final ProtoFileDownload.FileDownload.Selector selector, final int progress, final long userId, final RoomType roomType) {
+    public void onAvatarDownload(final String token, final long offset, final ProtoFileDownload.FileDownload.Selector selector, final int progress, final long userId, final RoomType roomType) {
         // empty
     }
