+++ app/src/main/java/com/iGap/adapter/items/AvatarItem.java
@@ -55,7 +55,7 @@ public class AvatarItem extends AbstractItem<AvatarItem, AvatarItem.ViewHolder>
     private void requestForAvatarFile(String token) {
         if (!AvatarsAdapter.hasFileRequested(token)) {
             AvatarsAdapter.requestsProgress.put(token, 0);
-            AvatarsAdapter.requestsOffset.put(token, 0);
+            AvatarsAdapter.requestsOffset.put(token, 0L);
 
             onRequestDownloadAvatar(0, 0);
         }
@@ -102,7 +102,7 @@ public class AvatarItem extends AbstractItem<AvatarItem, AvatarItem.ViewHolder>
     }
 
     @Override
-    public void onRequestDownloadAvatar(int offset, int progress) {
+    public void onRequestDownloadAvatar(long offset, int progress) {
         ProtoFileDownload.FileDownload.Selector selector =
                 ProtoFileDownload.FileDownload.Selector.FILE;
         final String fileName = avatar.getToken() + "_" + avatar.getName();
