+++ app/src/main/java/com/iGap/fragments/FragmentShowAvatars.java
@@ -199,7 +199,7 @@ public class FragmentShowAvatars extends Fragment implements OnFileDownloadRespo
     }
 
     @Override
-    public void onFileDownload(final String token, final int offset,
+    public void onFileDownload(final String token, final long offset,
                                final ProtoFileDownload.FileDownload.Selector selector, final int progress) {
         getActivity().runOnUiThread(new Runnable() {
             @Override
@@ -216,7 +216,7 @@ public class FragmentShowAvatars extends Fragment implements OnFileDownloadRespo
     }
 
     @Override
-    public void onAvatarDownload(String token, int offset,
+    public void onAvatarDownload(String token, long offset,
                                  ProtoFileDownload.FileDownload.Selector selector, int progress, long userId,
                                  RoomType roomType) {
         // empty
