+++ app/src/main/java/com/iGap/interfaces/IChatItemAvatar.java
@@ -4,5 +4,5 @@ package com.iGap.interfaces;
  * Created by Alireza Eskandarpour Shoferi (meNESS) on 9/28/2016.
  */
 public interface IChatItemAvatar {
-    void onRequestDownloadAvatar(int offset, int progress);
+    void onRequestDownloadAvatar(long offset, int progress);
 }
