+++ app/src/main/java/com/iGap/interfaces/OnFileDownloadResponse.java
@@ -7,10 +7,10 @@ import com.iGap.realm.enums.RoomType;
  * Created by Alireza Eskandarpour Shoferi (meNESS) on 9/28/2016.
  */
 public interface OnFileDownloadResponse {
-    void onFileDownload(String token, int offset, ProtoFileDownload.FileDownload.Selector selector,
+    void onFileDownload(String token, long offset, ProtoFileDownload.FileDownload.Selector selector,
                         int progress);
 
-    void onAvatarDownload(String token, int offset,
+    void onAvatarDownload(String token, long offset,
                           ProtoFileDownload.FileDownload.Selector selector, int progress, long userId,
                           RoomType roomType);
